Correctly type stream bar replicant in graphic
