+++ stream-bar/extension/StreamBarInfo.ts
@@ -1,0 +1,8 @@
+export interface StreamBarInfo {
+    lastBomb?: string;
+    lastSubscriber?: string;
+    lastTip?: string;
+    lastCheer?: string;
+    songName?: string;
+    artistName?: string;
+}
